Fix lint errors and fix some duplications
